==> notebooks/0_download_data.ipynb <==
import kagglehub

# Download latest version
path = kagglehub.dataset_download("yelp-dataset/yelp-dataset")

print("Path to dataset files:", path)

!mv /home/richardarcher/.cache/kagglehub/datasets/yelp-dataset ../data/raw
