==> notebooks/3b_offtheshelf_openai.ipynb <==
# import os
import polars as pl
import asyncio
from openai import AsyncOpenAI

# import re

import nest_asyncio
nest_asyncio.apply()

# client = AsyncOpenAI(api_key="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")
client = AsyncOpenAI(api_key="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")

async def fetch_prompt(sys_prompt: str, user_prompt: str):
    """Fetch a single chat completion asynchronously."""
    completion = await client.chat.completions.create(
        messages=[
            {"role": "system", "content": sys_prompt},
            {"role": "user", "content": user_prompt}
        ],
        model="gpt-4o",
    )
    return completion

async def main(lst_system_prompts, lst_prompts):
    # Create a list of tasks for all system/user prompt pairs
    tasks = [
        asyncio.create_task(fetch_prompt(sys_p, usr_p))
        for sys_p, usr_p in zip(lst_system_prompts, lst_prompts)
    ]

    # Run all tasks concurrently
    results = await asyncio.gather(*tasks)
    return results

def create_prompt(review):
    system_prompt = f"You read Yelp reviews and return a number (1, 2, 3, 4, or 5) that represents your besst guess of the number of star ratings that were given by that reviewer. Return just the number 1, 2, 3, 4, or 5, with no context, explanation, or special symbols."
    prompt = f"Here is the review to evaluate: [[[{review}]]]. Remember, you read Yelp reviews and return a number (1, 2, 3, 4, or 5) that represents your besst guess of the number of star ratings that were given by that reviewer. Return just the number 1, 2, 3, 4, or 5, with no context, explanation, or special symbols."
        
    return system_prompt, prompt

df_val = pl.read_csv("../data/1_train_test_split/df_validation.csv")

test_texts = df_val["text"].to_list()
test_labels = df_val["stars"].to_list()

lst_system_prompts, lst_prompts = [], []
for review in test_texts:
    system_prompt, prompt = create_prompt(review)
    lst_system_prompts.append(system_prompt)
    lst_prompts.append(prompt)

len(lst_system_prompts), len(lst_prompts)

chunk_size=105
lst_of_lsts_of_system_prompts = [lst_system_prompts[i:i + chunk_size] for i in range(0, len(lst_system_prompts), chunk_size)]
lst_of_lst_prompts = [lst_prompts[i:i + chunk_size] for i in range(0, len(lst_prompts), chunk_size)]

lst_of_results = []

print(f"GOING TO DO {len(lst_of_lsts_of_system_prompts)} MANY ITERATIONS")
ix=0
for lst_system_prompts, lst_prompts in zip(lst_of_lsts_of_system_prompts, lst_of_lst_prompts):
    ix = ix + 1
    print(f"NOW ON {ix=}")
    results = asyncio.run(main(lst_system_prompts, lst_prompts))
    lst_of_results.append(results)

results = []
for lst_ in lst_of_results:
    for i in lst_:
        results.append(i)

res = [result.choices[0].message.content for result in results]

len(res)

res_int = [int(i) for i in res]

right, total = 0, 0
for pred, actual in zip(res_int, test_labels):
    if pred==actual:
        right += 1
    total += 1

print(right/total)

df_val = df_val.with_columns(pl.Series(res_int).alias("4o_prediction"))

df_val.write_csv("../data/3_outputs/4o_predictions_for_eval_set.csv")
